Add IBTK::QuadratureCache.
This stores libMesh quadrature objects (things inheriting from
libMesh::QBase) in a map.
